OnHands
Omg I tried so many things to try to get hand tracking to work. All of the modifications I added are in the OnHands folder except for the new map. I created a new scene so that we can just switch scenes for testing and experimenting. Otherwise I messed with the OnHands script and tried with two new pawns. MetaXR plugin project settings are also set to controllers and hands, not just controllers.